refactor: move hooks from api folder
