+++ apps/web/src/components/auth/CreateOrganizationForm.tsx
@@ -8,7 +8,7 @@ import { IJwtPayload } from '@novu/shared';
 import { Button, Input } from '../../design-system';
 import { api } from '../../api/api.client';
 import { useAuthContext } from '../../store/authContext';
-import { useVercelIntegration } from '../../api/hooks/useVercelIntegration';
+import { useVercelIntegration } from '../../hooks/useVercelIntegration';
 import { useVercelParams } from '../../hooks/useVercelParams';
 import { ROUTES } from '../../constants/routes.enum';
