+++ apps/web/src/hooks/integrations/useActiveIntegrations.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getActiveIntegrations } from '../../integration';
+import { getActiveIntegrations } from '../../api/integration';
 
 export function useActiveIntegrations() {
   const { data, isLoading, refetch } = useQuery(['notificationsList'], getActiveIntegrations);
