+++ apps/web/src/hooks/integrations/useIntegrationLimit.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { ChannelTypeEnum } from '@novu/shared';
-import { getIntegrationLimit } from '../../integration';
+import { getIntegrationLimit } from '../../api/integration';
 import { useMemo } from 'react';
-import { IS_DOCKER_HOSTED } from '../../../config/index';
+import { IS_DOCKER_HOSTED } from '../../config/index';
 import { useActiveIntegrations } from './useActiveIntegrations';
 
 export function useIntegrationLimit(type: ChannelTypeEnum) {
