+++ apps/web/src/hooks/useNotificationGroup.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getNotificationGroups } from '../notification-groups';
+import { getNotificationGroups } from '../api/notification-groups';
 
 export function useNotificationGroup() {
   const { data, isLoading } = useQuery(['notificationGroups'], getNotificationGroups);
