+++ apps/web/src/hooks/useSubscribers.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { ISubscriber } from '@novu/shared';
 
-import { getSubscribersList } from '../subscribers';
-import { useEnvController } from '../../store/useEnvController';
+import { getSubscribersList } from '../api/subscribers';
+import { useEnvController } from '../store/useEnvController';
 
 export function useSubscribers(page = 0, limit = 10) {
   const { environment } = useEnvController();
