+++ apps/web/src/pages/quick-start/components/OnboardingSteps.tsx
@@ -8,10 +8,10 @@ import useStyles, { ActiveWrapper, StyledDescription, StyledTitle } from './Onbo
 import { Button, colors, Text } from '../../../design-system';
 import { CheckCircle } from '../../../design-system/icons';
 import { TriggerCard } from './TriggerCard';
-import { useTemplates } from '../../../api/hooks/useTemplates';
+import { useTemplates } from '../../../hooks/useTemplates';
 import { getActivityStats } from '../../../api/activity';
 import { ROUTES } from '../../../constants/routes.enum';
-import { useIntegrations } from '../../../api/hooks';
+import { useIntegrations } from '../../../hooks';
 
 export const OnboardingSteps = ({ onFinishedAll }: { onFinishedAll: () => void }) => {
   const navigate = useNavigate();
