+++ apps/web/src/pages/quick-start/components/TestNotificationTrigger.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { LoadingOverlay, Stack, useMantineTheme } from '@mantine/core';
 
 import { Button, colors, Text } from '../../../design-system';
-import { useTemplates } from '../../../api/hooks/useTemplates';
+import { useTemplates } from '../../../hooks/useTemplates';
 import { notificationTemplateName, OnBoardingAnalyticsEnum, onBoardingSubscriberId } from '../consts';
 import { errorMessage, successMessage } from '../../../utils/notifications';
 import { testTrigger } from '../../../api/notification-templates';
